+++ Backend/server.js
@@ -2,6 +2,7 @@ import express from "express"
 import authRoutes from './routes/authRoutes.js'
 import trendingRoutes from './routes/trendingSongRoutes.js'
 import searchRoutes from './routes/searchRoutes.js'
+import savedSongsRoutes from './routes/savedSongRoutes.js'
 const app = express()
 const PORT = 5200;
 app.use(express.json())
@@ -14,4 +15,5 @@ app.use((req, res, next) => {
 app.use('/auth', authRoutes)
 authRoutes.use('/login/trending', trendingRoutes)
 authRoutes.use('/search', searchRoutes)
+authRoutes.use('/favorites', savedSongsRoutes)
 app.listen(PORT, console.log(`Successful connection to ${PORT}`))
